Add files via upload
Added an analysis tab to add H (alpha/beta/gamma/delta) lines to the spectra. Like with the min/max tab, if you modify the plot after you add the lines, the lines will dissappear. 
Also changed the xlim/ylim entries on the gui to accept floats instead of integers.